Catch relevant error during callback phase
I am not sure under which circumstances the original
error condition yielding the CallbackError occurs. The
newly added error occurs when we try to exchange the
access token, but the client credentials are wrong.
Thus I assume that we'd receive the same for all errors
specified by the OAuth 2.0 spec.